update bashrecon.sh, added sftp support
Added built in sftp support. SFTP (when using a pub key) is much more secure than email, so I decided to make that the default means of transporting the logs. If you don't want/need it, remove the line 'phone_HOME' near the end. If you run this as a cron and do want to use sftp, than redirect the output to /dev/null so that it doesn't email you anyway.