Update: auto grab interfaces & delete temp files
Now the script automatically gets a list of all of the system's interfaces from ifconfig rather than blindly looking for 'wlan0' and 'eth0'. Also added a feature that wipes all of temporary files from the system, for stealth reasons. I figure if it's going to be used as an anti theft measure, than it would do to not leave any traces of the reconaissance on the target system. The script will first try to use secure-delete (srm), and than revert to normal delete (rm) if that fails.